refactor(#2407): SonarCloud new-code dup — BG 발사 후처리 persistBgFireResult로 추출
- bgPositionTrainFire.ts / undergroundConsensusFire.ts의 processLocationUpdate
  후처리(firedAlarms/ALARM_EVENT_KEY/BG_LAST_STATION_KEY/위젯 forward)가 동일 로직
  반복 — src/features/alarm/utils/bgFirePersist.ts의 persistBgFireResult로 추출
- 동작 무변경 — 순수 추출, 두 호출부 모두 동일 파라미터로 위임
